+++ examples/web/pitchassets/js/onload/load.js
@@ -7,6 +7,6 @@
 */
 jQuery(window).on('load', function(){
   jQuery(function(){ /* load after DOMREADY */
-    console.log('load $baseUrl');
+    console.log('load $baseUrl/');
   }).delay(1);
 });